Consolidated desk files; started building result_lookup
